% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glot_med}
\alias{glot_med}
\title{Glottolog 4.6 data: Agglomerated Endangerment Status}
\format{
A tibble with 8,345 rows and 18 variables:
\describe{
  \item{ID}{Primary key.}
  \item{Language_ID}{Reference to `languages$ID`.}
  \item{Parameter_ID}{Parameter ID.}
  \item{Value}{Parameter value.}
  \item{Code_ID}{Code ID.}
  \item{Comment}{Comment.}
  \item{Source}{Source reference.}
  \item{codeReference}{}
  \item{MED}{Most Extensive Description level.}
  \item{Name}{Name of the language.}
  \item{Macroarea}{Geographic macroarea.}
  \item{Latitude}{Latitude.}
  \item{Longitude}{Longitude.}
  \item{Glottocode}{Glottocode.}
  \item{ISO639P3code}{ISO639-3 code.}
  \item{Countries}{Countries where the language is spoken.}
  \item{Family_ID}{Family ID.}
  \item{Language_ID.y}{Language ID.}
}
}
\source{
\url{https://glottolog.org}
}
\usage{
glot_med
}
\description{
A tibble with the Most Extensive Description level (MED) in Glottolog 4.6.
}
\keyword{datasets}
