% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{Glottolog 4.6 data}
\format{
A tibble with 26285 rows and 12 variables:
\describe{
  \item{Language_ID}{Language ID.}
  \item{Glottocode}{Glottocode.}
  \item{Name}{Name of the language.}
  \item{Macroarea}{Geographic macroarea.}
  \item{Latitude}{Latitude.}
  \item{Longitude}{Longitude.}
  \item{Countries}{Countries where the language is spoken.}
  \item{Family_ID}{Family ID.}
  \item{AES_value}{Numeric value of the Aggregated Endangerment Status.}
  \item{AES}{Aggregated Endangerment Status.}
  \item{MED_value}{Numeric value of the Most Extensive Description types.}
  \item{MED}{Most Extensive Description type.}
}
}
\source{
\url{https://glottolog.org}
}
\usage{
glottolog
}
\description{
A tibble with data from Glottolog 4.6.
}
\keyword{datasets}
