% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottologR-package.R
\docType{package}
\name{glottologR-package}
\alias{glottologR}
\alias{glottologR-package}
\title{glottologR: Data from Glottolog 4.6}
\description{
It provides data from the Glottolog database (version 4.6).
}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID})

Other contributors:
\itemize{
  \item Harald Hammarström (Authors of Glottolog) [copyright holder, data contributor]
  \item Robert Forkel (Authors of Glottolog) [copyright holder, data contributor]
  \item Martin Haspelmath (Authors of Glottolog) [copyright holder, data contributor]
  \item Sebastian Bank (Authors of Glottolog) [copyright holder, data contributor]
}

}
\keyword{internal}
