% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{languages}
\alias{languages}
\title{Glottolog 4.6 data: "languages.csv"}
\format{
A tibble with 26,285 rows and 10 variables:
\describe{
  \item{ID}{Primary key.}
  \item{Name}{Name of the language.}
  \item{Macroarea}{Geographic macroarea.}
  \item{Latitude}{Latitude.}
  \item{Longitude}{Longitude.}
  \item{Glottocode}{Glottocode.}
  \item{ISO639P3code}{ISO639-3 code.}
  \item{Countries}{Countries where the language is spoken.}
  \item{Family_ID}{Family ID.}
  \item{Language_ID}{Language ID.}
}
}
\source{
\url{https://glottolog.org}
}
\usage{
languages
}
\description{
A tibble with data from "languages.csv" in Glottolog 4.6.
}
\keyword{datasets}
