% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{values}
\alias{values}
\title{Glottolog 4.6 data: "values.csv"}
\format{
A tibble with 131,048 rows and 8 variables:
\describe{
  \item{ID}{Primary key.}
  \item{Language_ID}{Reference to `languages$ID`.}
  \item{Parameter_ID}{Parameter ID.}
  \item{Value}{Parameter value.}
  \item{Code_ID}{Code ID.}
  \item{Comment}{Comment.}
  \item{Source}{Source reference.}
  \item{codeReference}{}
}
}
\source{
\url{https://glottolog.org}
}
\usage{
values
}
\description{
A tibble with data from "values.csv" in Glottolog 4.6.
}
\keyword{datasets}
